requirements.txt
pandas
numpy
setuptools
google-cloud-storage
pyyaml
matplotlib
tensorflow
wordcloud
scikit-learn
comet-ml
dvc